Fix forEachLayerAtPixel for the canvas renderer
